python-web-app/pyproject.toml
[tool.poetry]
name = "python-web-app"
version = "0.1.0"
description = ""
authors = ["ryota0051 <re0061vh@gmail.com>"]

[tool.poetry.dependencies]
python = "3.9.13"
fastapi = "0.81.0"
uvicorn = "0.18.3"
python-multipart = "0.0.5"
Pillow = "9.2.0"
onnxruntime = "1.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
